refactor: UIColor Extension 파일 생성 이미지 파일 이름 상수 파일에 생성 validInputLength -> idValidInputLength 수정
