+++ LoginProfile/Common/Constants.swift
@@ -10,12 +10,15 @@ import UIKit
 struct Constant {
     static let yellow = 0xFBE64C
     static let lightgray = 0xF6F6F6
-    static let validInputLength = 5
+    static let idValidInputLength = 5
     static let menuLabelFontSize = CGFloat(17)
     static let bottomMenuColor = 0xFFFFFF
     static let chatMenuLabelText = "1:1 채팅"
     static let callMenuLabelText = "통화하기"
     static let kakaoStoryMenuLabelText = "카카오스토리"
+    static let chatImageName = "btn_bubble"
+    static let callImageName = "btn_phone"
+    static let kakaoStoryImageName = "btn_quote"
 }
