+++ LoginProfile/Common/UIColor+RGB.swift
@@ -1,0 +1,27 @@
+//
+//  UIColor+RGB.swift
+//  LoginProfile
+//
+//  Created by 이승주 on 2022/01/07.
+//
+
+import UIKit
+
+extension UIColor {
+    convenience init(red: Int, green: Int, blue: Int, alpha: Int = 0xFF) {
+        self.init(
+            red: CGFloat(red) / 255.0,
+            green: CGFloat(green) / 255.0,
+            blue: CGFloat(blue) / 255.0,
+            alpha: CGFloat(alpha) / 255.0
+        )
+    }
+
+    convenience init(rgb: Int) {
+        self.init(
+           red: (rgb >> 16) & 0xFF,
+           green: (rgb >> 8) & 0xFF,
+           blue: rgb & 0xFF
+        )
+    }
+}
