+++ LoginProfile/ViewControllers/LoginViewController.swift
@@ -1,5 +1,5 @@
 //
-//  LoginProfile - ViewController.swift
+//  LoginProfile - LoginViewController.swift
 //  Created by yagom. 
 //  Copyright © yagom. All rights reserved.
 // 
@@ -26,7 +26,7 @@ class LoginViewController: UIViewController {
             return
         }
         
-        if input.count >= Constant.validInputLength {
+        if input.count >= Constant.idValidInputLength {
             loginButton.isEnabled = true
         } else {
             loginButton.isEnabled = false
@@ -64,24 +64,5 @@ extension LoginViewController: UITextFieldDelegate {
         }
         passwordTextField.resignFirstResponder()
         return true
-    }
-}
-
-extension UIColor {
-    convenience init(red: Int, green: Int, blue: Int, alpha: Int = 0xFF) {
-        self.init(
-            red: CGFloat(red) / 255.0,
-            green: CGFloat(green) / 255.0,
-            blue: CGFloat(blue) / 255.0,
-            alpha: CGFloat(alpha) / 255.0
-        )
-    }
-
-    convenience init(rgb: Int) {
-        self.init(
-           red: (rgb >> 16) & 0xFF,
-           green: (rgb >> 8) & 0xFF,
-           blue: rgb & 0xFF
-        )
     }
 }
