+++ LoginProfile/Views/ProfileBottomMenuBar.swift
@@ -18,19 +18,19 @@ class ProfileBottomMenuBar: BaseStackView {
     }
     
     private lazy var chatMenu = ProfileMenu(
-        menuImage: UIImage(named: "btn_bubble"),
+        menuImage: UIImage(named: Constant.chatImageName),
         menuLabeltext: Constant.chatMenuLabelText,
         spacing: CGFloat(17.58)
     )
     
     private lazy var callMenu = ProfileMenu(
-        menuImage: UIImage(named: "btn_phone"),
+        menuImage: UIImage(named: Constant.callImageName),
         menuLabeltext: Constant.callMenuLabelText,
         spacing: CGFloat(17.58)
     )
     
     private lazy var kakaoStoryMenu = ProfileMenu(
-        menuImage: UIImage(named: "btn_quote"),
+        menuImage: UIImage(named: Constant.kakaoStoryImageName),
         menuLabeltext: Constant.kakaoStoryMenuLabelText,
         spacing: CGFloat(17.58)
     )
